Remove default node server
