+++ pkg/azurelustre/azurelustre.go
@@ -131,8 +131,8 @@ type LustreSkuValue struct {
 type Driver struct {
 	csicommon.CSIDriver
 	csicommon.DefaultIdentityServer
-	csicommon.DefaultNodeServer
 	csi.UnimplementedControllerServer
+	csi.UnimplementedNodeServer
 	// enableAzureLustreMockMount is only for testing, DO NOT set as true in non-testing scenario
 	enableAzureLustreMockMount bool
 	// enableAzureLustreMockDynProv is only for testing, DO NOT set as true in non-testing scenario
@@ -176,7 +176,6 @@ func NewDriver(options *DriverOptions) *Driver {
 	d.NodeID = options.NodeID
 
 	d.DefaultIdentityServer.Driver = &d.CSIDriver
-	d.DefaultNodeServer.Driver = &d.CSIDriver
 
 	ctx := context.Background()
